Implementation de certaines fonctions generales

+++ Symulation_Ecosysteme/Symulation_Ecosysteme/ClassesSimulateur/Coordonne.cpp
@@ -20,7 +20,7 @@ double Coordonne::getX() {
 /**
  * @return int
  */
-int Coordonne::getY() {
+double Coordonne::getY() {
     return 0;
 }
