+++ Symulation_Ecosysteme/Symulation_Ecosysteme/ClassesSimulateur/Coordonne.h
@@ -11,7 +11,7 @@ public:
     
 double getX();
     
-int getY();
+double getY();
     
 void setX();
